Create grid and initial cards
